+++ ufs2arco/__init__.py
@@ -1,7 +1,8 @@
 __version__ = "0.1.2"
 
 from .cice6dataset import CICE6Dataset
-from .fv3dataset import FV3Dataset, Layers2Pressure
+from .fv3dataset import FV3Dataset
+from .layers2pressure import Layers2Pressure
 from .mom6dataset import MOM6Dataset
 from .regrid import MOM6Regridder, CICE6Regridder
 from .timer import Timer
